requirements.txt
amqp==5.2.0
astroid==3.0.3
async-timeout==4.0.3
billiard==4.2.0
black==24.2.0
blinker==1.7.0
celery==5.3.6
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.8
distlib==0.3.8
exceptiongroup==1.2.0
filelock==3.13.1
flake8==7.0.0
Flask==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
identify==2.5.35
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pyflakes==3.2.0
pylint==3.0.3
pytest==8.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
SQLAlchemy==2.0.27
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
youtube-dl==2021.12.17
